requirements.txt
python-telegram-bot
python-dotenv
datapizza-ai
datapizza-ai-parsers-docling
datapizza-ai-clients-google
datapizza-ai-embedders-fastembedder 
fastembed 
PyYAML

# Dev Dependencies
pytest
pytest-asyncio
pytest-mock
scikit-learn
